move zmq interface classes to control_interface
